remove lang
