+++ package.json
@@ -54,7 +54,7 @@
     "grunt-ts": "^5.5.1",
     "grunt-tslint": "^3.1.0",
     "grunt-typings": ">=0.1.5",
-    "intern": "^3.4.1",
+    "intern": "3.3.2",
     "istanbul": "^0.4.5",
     "jsdom": "^9.5.0",
     "postcss-modules": "^0.5.1",
