--- src/examples/flexbox.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<style>
-		@import url('../../_build/src/themes/basic/basic.css');
-	</style>
-</head>
-<body class="basic">
-	<dojo-container-layout class="horizontal">
-		<dojo-panel-tabbed>
-			<ul>
-				<li class="active">
-					<div class="tab-label">Panel One</div>
-					<div class="tab-close">X</div>
-				</li>
-				<li>
-					<div class="tab-label">Panel Two</div>
-					<div class="tab-close">X</div>
-				</li>
-				<li>
-					<div class="tab-label">Panel Three</div>
-					<div class="tab-close">X</div>
-				</li>
-			</ul>
-			<div class="panels">
-				<dojo-panel class="visible">foo</dojo-panel>
-				<dojo-panel>bar</dojo-panel>
-			</div>
-		</dojo-panel-tabbed>
-	</dojo-container-layout>
-</body>
-</html>
