--- src/examples/index.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<link rel="stylesheet" type="text/css" href="../themes/basic/basic.css">
-	<title>dojo-widget Examples</title>
-	<style>
-		body {
-			padding-bottom: 100px;
-		}
-
-		dojo-panel-tabbed {
-			max-width: 1000px;
-		}
-
-		dojo-panel-tabbed > div:nth-child(2) {
-			min-height: 300px;
-		}
-
-		.buttons {
-			max-width: 320px;
-		}
-
-		.buttons button {
-			margin-bottom: 30px;
-		}
-
-		.buttons button:nth-child(2n) {
-			margin-right: 20px;
-		}
-
-		.buttons button:nth-child(5),
-		.buttons button:nth-child(6) {
-			margin-bottom: 0;
-		}
-	</style>
-</head>
-<body class="basic">
-	<!-- Loading the dojo-loader (minified) -->
-	<script src="../../../node_modules/dojo-loader/loader.min.js"></script>
-	<!-- Loading the boot script -->
-	<script src="./boot.js"></script>
-</body>
-</html>
